Makes cap upgrades remember their limits
This change makes capacitor banks remember the limits which were set to them before being expanded.
This way, a single block capacitor bank with an IO of 100MJ/t and user-specified output of 50 MJ/t,
when expanded with another bank next to it will become a bank with IO of 200MJ/t and user-specified
output of... 50MJ/t! However, if you have a bank which has maximum input or output, that will expand.
Since input and output is handled seperately, you can have an output limit while still expanding the
input limit.