Important: Please see related issue #580, as this modifies code added to prevent that situation.
Activating Extract mode causes the receivedTicks collection
to be created. This collection is never removed and is not
updated except when receiving energy, which is not done when
Insert or In/Out is active.

This change overrides AbstractPowerConduit.setConnectionMode() to
handle clearing the extra state added by Extract mode, which allows
Insert or In/Out to be used on cap banks after Extract has been selected
once.

This issue was initially reported as Insert mode not working when initially
set. After the server was restarted the conduit would then work properly.
The problem was actually the scrolling past Extract mode on the way to Insert.