Different Item-Place-Logic
this fixes CrazyPants/EnderIO/#1648

the original way caused problems with items that relied on
player.inventory.currentItem, like cables from AE2 - they would set the
.currentItem-Stack to null when you placed the last cable, which in our
case would kill the pick.

Also, we’re now using the bult-in “can we place the item there?” logic
and can therefore remove canPlaceBlockOnRightClick.

It’s still not perfect though, because it doesn’t work with buckets…
see the source-code-comments. A possible solution would be to implement
a custom message that we send to ourself to switch back to the original
item slot, but I’ll just go with “you’re holding the bucket wrong”.