Fixes for the Crafter
* Fixes #2038. Recipes that return different amounts based on input item
NBT or phase of the moon will result in the correct amount now.
* Fix for the same case: Result itemstacks will never be too big.
* Fix: Recipes that return different NBT data based upon input item NBT
or phase of the moon will be handled (nearly) correctly. Remaining error
is that there will be an output queue instead of the crafting being
paused. Not worth to fix.
* Fix: Overflow items are now merged into the result stack correctly.
* Improvement: Container items should now be handled correctly no matter
what kind of funny things a mod does with its items.