Modified Molten Ender Teleport
Modified the teleportation for molten ender to trigger an event as if an ender pearl has been thrown for the teleportation of a mob/player. I did this mainly to work with the Dark Utils Ender Tether as it adjusts the teleport destination based on the EnderTeleportEvent. Teleport still behaves as normal and random if there is nothing that intercepts the event.