Sould Binder Temp Fix
Temperary fix for the soul binder (can now only craft if the output slot is empty (prevents soul vial (bat) + broken spawner from creating a slime spawner if a slime spawner is already in that slot)).  canStartNextTask (should) check if the recipe output is equal to the item in the output slot or if the slot is empty.