Update TileConduitBundle.java
Greatly increase Client Side ItemConduit performance. (20% fps increase if have 100 - 200 item conduit in render area, these are lag-spikes.)

EnderIO Conduits are very laggy, partly because Redstone/fluid/gas Conduits causing RenderChunks re-render back and forth. But if there is only one ItemConduit in the bundle, it is pretty sure that we only need to re-render ItemConduit near the player , because only player's behavior range (5) can change the shape of ItemConduit. 
But if there is another conduit in the bundle, the re-render is not sure because Redstone conduit in the bundle might needs to re-render in any time. So in this case I left to default.